"set pm3d border retrace" will use quadrangle color for the border
In principle this should not be necessary, but it mitigates an issue
that some output modes suffer from antialiasing glitches between
adjacent pm3d quadrangles.  This is particularly common for pdf
viewers with antialiasing enabled (note that this is really a problem
in the viewer, not in the pdf file).  Retracing the border in the
same color as the fill hides this antialiasing glitch, at the cost
of a larger output file.

This essentially adds a documented keyword "retrace" that triggers
the previously undocumented effect of "set pm3d border lt -6".