revised policy for enabling palette colors in 2D
Previous policy: Check plot components (everything in the "plot"
command) for use of palette color, check objects and labels for use
of palette color.  If present, enable the colorbox.  If not,
palette colors are not usable.

Problem: Other elements like arrows, grid lines, textbox borders,
etc could conceivably be specified to use palette colors even if the
plot components do not.  They would all be rendered in black if the
palette is not enabled.

New policy:  Check the plot components for palette use; enable the
colorbox if present.  Do not bother to check objects, labels, etc.
Regardless of whether or not the colorbox is enabled, palette colors
are used if requested.