New option "set style histogram nokeyseparators"
If there are multiple histograms in a single plot command, the
program by default adds a blank line to the key so that the titles
belonging to one histogram are separated from titles belonging to
the next histogram.  This leads to unwanted blank space in the key
if the individual histogram components do not have separate titles.
In this case or for any other reason you want to suppress the extra
blank lines, you can now use "set style histogram nokeyseparators".

Bug #2473