Rearrange "struct coordinate" to avoid a hole due to packing
Old:	struct coordinate {
		enum coord_type type;
		coordval x,y,z,ylow,yhigh,xlow,xhigh;
	}
New:	struct coordinate {
		coordval x,y,z,ylow,yhigh,xlow,xhigh;
		enum coord_type type;
		int extra;
	}

The compiler (at least current gcc and clang) inserted 4 or 6 bytes of
unused spaced after "type" depending on whether the enum was
treated as a short or an int.
I suppose this is to guarantee 8-byte alignment of the full structure.

- This caused error warnings from valgrind about uninitialized
  memory when comparing two structures.
- If there are 4+ bytes are unused space available, we might as well
  get some use out of them by adding an explicit "int extra" to the structure.

For gcc and clang both the old and new structure definitions yield
	sizeof(struct coordinate) == 64

If there is some platform for which the old sizeof() was 58 or 60,
at worst there is a size penalty for huge numbers of data points.

To disable:
	./configure  --without-extra-coordinate