Allow unicode characters outside the BMP
The documentation claims support for Unicode code points via
4 or 5 character hexadecimal representation.
For example "\U+1F319" (crescent moon) requires 5 hexidecimal characters.
This worked when embedded in a larger string, but use as a single character
was broken because the 5th character, if any, was lost in
encoding.c:truncate_to_one_utf_char().