Handle additional cross-references in Japanese documentation
Not all of the cross reference sites use 以下参照:
some simply embed 参照 in the preceding sentence.
This creates one false-positive (same word used in a normal sentence
that just happens to have a highlighted phrase later in the text).

Do not accept `$word` (anything starting with $) as an index entry
or cross-reference because the $ messes up TeX parsing.