add infrastructure for plot filters
The goal is to move some of the current smoothing operations that
have little to do with any intuitive notion of "smoothing" into a
new category of data operations called "filters".  Typically a
filter will select only a subset of the points in the input stream
(e.g. "convexhull", "mask") or rearrange and restructure the points
("bins").

A distinguishing feature is that filters are applied immediately after
reading in data, prior to any smoothing.

Filter selection is tracked in a new field of the plot/splot
header structure plot->plot_filter, parallel to the existing options
tracked in plot->plot_smooth.

In principle you can apply both a filter and a smoothing operation.
Most combinations make no sense, but for example
"plot FOO convexhull smooth path" is essentially the same as
"plot FOO smooth convexhull", which was the original syntax.
We don't currently have smoothing options for image plots, but one could
imagine a similar combination "splot FOO mask smooth lancosz with rgbimage"

This initial work
- makes "convexhull" a filter rather than a smooth option.
- makes "bins" a filter rather than a smooth option.
- makes "mask" a filter.