block: support block octants
The 2x4 block octant characters offer a higher resolution and yield
roughly square pixels when taking the aspect ratio of typical screen fonts
into account. They are not part of Unicode, but available in some fonts
like "FairFax HD".

While most of this is a trivial extension of the code, the optimization of
fully occupied character cells requires a table with 141 bit patterns.
These are just permutations of 7, 6, 5, and 4 bits set out of 8 total,
which in principle could be generated on the fly.