block: revised color optimization for full cells
For full character cells with more than a single color, we use fore-
and background colors to improve the presentation. The code now searches
for the optimum combination of pixels to average by trying out all
combinations. This brute-force method is acceptable due to the
small number of pixels in a cell and in total.