defer evaluation of 2D plot title until after filters and smoothing
In version 5 evaluation of a plot title that is a string-valued
expression is deferred until after data is read in, so that it can
refer to things like column headers that are no known before then.
Now we defer it even longer, until after filter, smoothing, and
image processing operations, so that any stats or other info
generated there can be captured in the plot title.