docs:  replace "make html"
I have given up getting the old pathway of *.doc -> *.tex -> *.html
to work again.  This commit takes the easier path of starting with
the windows + wxt help system code in windows/doc2html and modifying
it to produce a set of html files that can stand on their own as a URL.

I chose to use svg for embedded sample plots and other figures.