html docs:  more figures and hyperlinks
Provide two figures as png images because they are not available as svg
	figure_missing
	figure_static_d20

Modify figure to display text corrctly in svg
	demo/unicode.dem

Provide hyperlink to demos
	hypertext
	unicode

Move all figure points to _after_ the section title, not before. E.g.
	3 Some new topic
	Ffigure_new_topic
This make the TeX output more prone to leaving a section title hanging
on the previous page, but if the figure insertion comes first then it
ends up on the wrong HTML page, which is worse.