Consistent support of alpha channel in "with image"
[Correction and partial reversion for dd12a0f9]

There is an internal fillstyle option FS_DEFAULT that was never well
defined.  The basic intent was to handle different defaults for different
object types, so that some could default to solid and others to empty.
It was never intended to apply to styles like "with image".
Also, to achieve consistency across the board it should be interpreted
in the core code rather than letting different drivers do different
things when they see FS_DEFAULT. Arguably they should never see it at all
because the core code has already replaced it with something more specific.

Change the image processing code to set FS_OPAQUE rather than FS_DEFAULT.
If there it uses a named palette with alpha channel, switch this to
FS_TRANSPARENT_SOLID + alpha.

This change allows images to be modified by alpha values in a named
palette in both 2D and 3D plots, e.g.
    "with image pixels fc palette NAMEDPALETTE"

Note that the "pixels" keyword is needed because the generic code
supporting transparency is only invoked when each pixel of the image
is drawn separately.  While the image support in various underlying
graphics libraries might in principle handle an alpha channel,
we don't rely on that.

Bug #2503