docs:  work around glitch in cross-indexing hyperlinks
The doc->tex->pdf pathway automatically capitalizes subsection
titles and thus doesn't care whether that word in gnuplot.doc was
upper or lower case to start with.  However the doc->html->...
pathway is case sensitive.  Some of the level 3 subsection references
were wrong or missing because the keyword was capitalized in the source.
This really wants a fix that deals with the larger problem,
but for now just change the gnuplot.doc so that the problematic entries
are lower case.