modify the checks for "missing" data
In version 5.4 the program caught the "missing" data flag when that column,
let's say column N, was referred to directly in a using specification as
N ($N) or (func($N)).  It did not catch references to (column(N)),
which should have been added by the fix for Bug #2042 (commit b8304eaf).
These direct checks are used to avoid evaluation of an expression that
depends on a missing value, forestalling floating point or other errors.

This patch modifies the checks for missing data.
1) There is now a pre-evaluation check for (column(N)).
2) If a missing value is encountered during evaluation of a using
   specifier¸ presumably after being missed by the direct checks,
   an internal flag is set so that the input is treated as missing just
   as in the direct test case.  This flag replaces use of a nonsense
   complex return value {NaN, DF_MISSING}.

The behavior is now described more completely in the documentation.