convex hull: fix clipping; add infrastructure for outlier rejection
When a point went out of range (outside the plot borders) it was
no longer considered as a possible member of the hull set.
This was wrong.  The hull should stay the same even if part of it
is now clipped.

Refactor the code so that the (new) tests for type == EXCLUDEDRANGE
can also be used to signal points that are outliers, and hence not
to be used when finding the hull.  However at this point there is
no user command to define the criteria for flagging outliers.