New filter operations delaunay_triangulation and concave_hull
Configuration option
	./configure --with-chi-shapes

Filter option for plot command
	chi_length = 0.5
	plot FOO using 1:2 concavehull ...

The routine delaunay_triangulation(plot) is implemented as a first step
in the construction of χ-shapes, which are concave hulls constructed by
selective removal of triangles from the Delaunay tesselation.
It may of course also be useful for other purposes.
Output:
    struct triangle *good_triangles
	A linked list of triangles in local storage whose vertices
	are indices into the array of data points in plot->points.
    struct t_edge bounding_edges[n_bounding_edges]
	An array of edges in local storage.
	Edge endpoints are indices into plot->points.
Side effects:
    points are sorted on x
    points lying on the convex hull are flagged
	point->extra = HULL_POINT

The routine delaunay_triangulation() implements the Bowyer-Watson algorithm
	A Bowyer (1981) Computer Journal 24:162.
	DF Watson (1981) Computer Journal 24:167.
except that it uses linked lists to hold triangles.  It would be cleaner
to take advantage of a generic list handling library; alas there is no
standard one for C.  We could use the list routines in Qt, but then this
feature would only be available if the qt terminal is configured in.

The routine concave_hull() implements contruction of a χ-shape (chi shape)
starting from the delaunay triangulation.  χ-shapes are as defined by
	M Duckham, L Kulik, M Worboys, and A Galton (2008)
	Pattern Recognition 41: 3224_3236.

Line segments on the perimeter of the tesselation, initially the convex hull,
are sorted by length.   The perimeter is then modified by iteratively
removing the longest segment subject to two criteria:
1) The length is greater than a pre-selected characteristic length (chi_length)
2) Removal of the segment does not cause the enclosed area to become
   disconnected or to hinge about a single vertex
Iteration stops when no such segment remains.
If no segment of the initial perimeter is longer than the
characteristic length, the result is the convex hull of the points.
As the characteristic length becomes shorter, more and more segments are
removed.  I.e. the "excavations" into the interior become larger and
the hull becomes more obviously concave.

Sorting the input points prior to triangulation enables a modification
to the basic algorithm that improves the run time for N points from O(N^2)
to approximately O(NlogN).

The characteristic length of the χ-shape in the most recent plot
is saved in GPVAL_CHI_LENGTH.

Notes:

- If chi_length is defined when the concave hull is requested, it is used
  as the characteristic length parameter to generate a χ-shape.

- If chi_length is not defined, the program sets it to
  chi_shape_default_length * max(edgelengths of convex hull)

- chi_shape_default_length is set to 0.6
  The undocumented command "set chi_shape fraction <value>" changes it.

- The Duckham paper recommended instead defaulting to a [small] fraction
  of the average edge length of the full Delaunay triangulation.
  That worked terribly for me; I think because they had a very different
  purpose and type of data in mind.

- If there is color information for the original set of points, keep
  it with the point through triangulation.  The color of one vertex is
  then inherited by each triangle created by save_delaunay_triangles.
    This begins to approximate what users may expect from triangulation,
  although gnuplot provides no way of smoothly blending the three vertex
  colors of a triangle to color each pixel separately.
  Because of that, I think dgrid3d does a better job.