Fix:   contour binary data "with labels"
It makes no sense to by default try to read two binary entries
(z + label text) from a binary file for each point on the
contoured surface.  The label will be generated automatically
by the program; it isn't stored as a data value.
Change the default number of binary entries per point to 1.