Allow variable point style and type in "plot ... with yerrorbars"
The documentation for plot style yerrorbars says "additional input
column[s ...] may be used to provide information such as variable
point color."  This implied that point color was not the only
property that could be marked "variable", but in fact yes it was
the only one.  Now we handle also variable point size and point type.

NB: There is a limitation that variable point type data cannot include
character strings, because the data field used to hold PT_CHAR
would conflict with the fields that hold ylow/yhigh for the error bar.