Allow 2D plots to handle arrow styles that include "lc rgb variable"
This means 2D plot styles VECTOR and ARROWS can accept up to 6
columns of data in the 'using' speciier.
	x y dx dy {variable arrowstyle} {variable color}
Before this tweak
1) the program would accept 5 columns maximum.
2) you could put "lc variable" in an arrow style but it would not
   be used in the plot because check_for_variable_color did not
   know to look for it.

Note that plotting with "arrowstyle variable" now always creates a
varcolor[] array in the plot structure, but the contents are used
only if the arrow style taken from data column 5 itself specifies
"lc variable" or "lc rgb variable".