Allow 3D plots to handle arrow styles that include "lc rgb variable"
    splot $DATA using ... with vector arrowstyle variable

If the arrowstyle (variable or not) specifies "lc variable" or
"lc rgb variable" then an additional column of input data must be
specified in the `using` clause.  This brings the total number
of input columns to 7 (constant arrowstyle with variable color)
or 8 (variable arrowstyle, some or all of which use variable color).