change: make "too many" columns in the using spec non-fatal
The program has evolved to support many optional data-dependent
properties, e.g. variable color, pointsize, arrowstyle, ...
and it turns out that some are hard to recognize in advance
(e.g. variable color inside a variable arrow style).
That is to say, sometimes the program mis-counts or fails to
predict the number of columns needed.

Change the code so that if this happens the user gets a warning
rather than a hard error.