Consistent ordering of columns for "plot ... ps variable pt variable"
The documented ordering of columns for variable data properties is
      plot DATA using x:y:pointsize:pointtype:color

This was correctly implemented for 2D plots in version 5.2 and for
both 2D and 3D plots "with points" in versions 5.2 and 5.4.
However 5.4 inadvertently reversed the ordering for 2D plots
"with points", "with linespoints", and "with labels" when both
ps and pt are variable.

Fix this so that 2D plots are consistent with equicalent 3D plots and
with the previous (v5.2) behavior.  Unfortunately this may affect
scripts written for 5.4.0 - 5.4.3, but long-term internal consistency
and backwards consistency seems to outweigh this.

This change also fixes a glitch that both 2D and 3D plots
"with labels point pointtype variable".  A data value of N actually
produce a label with point type N+1.

Bug #2524