improved stability of zoom_around_mouse()
If the rescaling from a '+' or '-' mousing hot key operation
causes the y axis tic labels to gain or lose a decimal point,
the position of the left plot border jumps and the nominal x coordinate
of the current mouse position jumps with it.  We now try to detect and
adjust for this jump in the simplest (linear x) case.

Bug #1375