Fix discrete gradient palette colorbox for terminals x11 cgm fig gd pstricks
The discrete gradient palette (DGP) introduced in commits
4ba95e0e 29f60657 5989ba41
did not generate a colorbox properly on some terminals.

For colorbox rendering in the case of DGP, the information in
sm_palette.gradient is used directly. On the other hand, some terminals that
define their own colormaps (x11, cgm, fig, gd, pstricks) generate smooth
colormaps that are incompatible with DGP, so in some cases the colorbox is not
drawn correctly. For other terminals this problem does not occur,
since the conversion from gray to RGB can be done directly using
rgb255maxcolors_from_gray, etc.

For the x11, cgm, fig, gd, and pstricks terminals, we now modify the process
of specifying palette colors from a gray value to handle discrete gradient
palettes separately.

To that end, the following changes were made

- definition of macro CHECK_SMPAL_IS_DISCRETE_GRADIENT to check if palette is DGP or not
- change make_palette() function to pass the original sm_palette.gradient
  content to the terminal instead of smooth colormap in case of DGP
- Separate the behavior of the set_color() function for each terminal into
  DGP and non-DGP cases
- changed to pass sm_palette.gradient_type in IPC between gnuplot and gplt_x11
- introduction of function index_from_gray() which refers to sm_palette.gradient
  and returns the index value corresponding to the gray value
- modify the call to index_from_gray in gplt_x11 to deal with the fact that the
  precision of sm_palette.gradient[N].pos is limited to "%6.4f" in IPC