More consistent bookkeeping for discrete gradient palette
This change sets sm_palette.gradient_type explicitly in "set_palette()"
and calls "check_palette_gradient_type()" if necessary.
Also, in situations where the palette is reset to initial state,
"sm_palette.gradient_type" is set in the "init_color()".
In this way, after calling command "set palette ...", we will know if the
palette is discrete or smooth (or mixed).