set xtics <incr> {time units}
E.g.
	set xtics time 5 years
	set xtics time "01-Jan-2001", 1 month, "01-Jan-2002"

Allow tic spacing to be specified in minutes/days/weeks/...
Non-integral values are accepted but probably not useful.

This makes the time interval easier to specify, but it doesn't
change the annoying fact that not all months have the same length.
A more accurate implementation might place tics along a time axis
sequentially by incrementing the relevant field in struct tm,
as done in gen_time_minitics().