hpgl:  Add "fontscale" terminal option
This option was requested to support a 1980s era HP 7475a 2-pen plotter.
The hpgl driver set a fixed character size that was proportional to the
plot area.  The size presumeably was chosen for large-bed plotters.
The 7475a uses letter size paper, and the characters come out very small.
	set term hpgl 2 fontscale 4.0
scales the characters back up to something usable on this plotter.