canvas terminal:  tweak javascript hypertext support
When hypertext anchor points were close together, the previous link
could still be active when a new one was triggered.  In this case the
previous text box was not erased and the plot area could gradually
accumulate lots of stale hypertext boxes. Now we clear all previous
boxes before drawing a new one.