Move function plot bookkeeping into the plot header
Previously all function plots used a single global data structure
command.c:
	struct udft_entry plot_func
This meant that for multiple functions in a single plot command,
each one overwrote the specification of the previous plot.
That works fine if everything is handled in a single pass through the
plot command, but if you want to go back afterwards to inspect individual
function plots the information is no longer around.

Now we move this into the plot header so that it can be accessed as
plot->plot_function.  This obviates the need for a global structure.