configuration option --enable-watchpoints
Conditional code framework to support development of EXPERIMENTAL
features based on actions triggered by the action of plotting.

The concept at this point is to attach watchpoints to individual plots,
e.g.
	plot FOO with lines watch y=100
that will trigger an action or store a value whenever the plot it
is attached to satisfies the target condition.  In this case it would
trigger whenever the line being plotted crossed the target value y=100.

Another example
	set style watch labels point pt 6 boxed
	plot for [i=1:N] file(N) watch mouse
This will track moused motion dynamically, displayed a label attached
to each of the N plotted lines at the point corresponding to the
current mouse position. Typically the label would report the y value.