New feature set: watchpoints
This patch set introduces support for a "watch" variant of the plot
style "with lines".  The code is EXPERIMENTAL, and conditionally
compiled under the control of configuration option --enable-watchpoints.

When executing a command "plot ... watch <target>" each line segment of
the resulting plot is checked to see if it satisfies one of the watchpoint
conditions.  If so, the [x,y] coordinates that satisfy the condition are
considered a "hit".

These targets are supported:
     x = <value>
     y = <value>
     z = <value>
     F(x,y) = <value>
     mouse
Targets z and F(x,y) are roughly equivalent in functionality, but providing
the function in principle allows reevaluation during interpolation
so that the reported hit coordinates are more accurate (not yet implemented).

If data smoothing options are in effect, the interpolation is done
along a line segment from the smoothed curve, _not_ interpolation between
two data points.

"plot FOO watch mouse" triggers when the plot crosses the current mouse
x position. This allows dynamic readout of the plot y values as the mouse moves.
It is ignored if mousing is not active in the current terminal.

Examples:
     set xrange [-10:10]
     plot x**2 with lines watch x=1.2 watch y=81.0
     print WATCH_1
         [{1.2,1.44}]
     print WATCH_2
         [{-9.0,81.0},{9.0,81.0}]

     # plot function together with previously found hits
     plot x**2 with lines, WATCH_2 using 2:3 with points

     # plot function with hits labeled as they are found
     set style watchpoints labels
     plot x**2 with lines watch x=1.2 watch y=81.0