Track columnheaders of multiple data blocks in a single file
"set datafile columnheaders" specifies that each data block in
a file begins with a line of columnheaders rather than data.
The code correctly dealt with this if you accessed any single
block in the file:
	plot foo index i using 1:2:(columnheader(2)) with labels
However if the entire file was processed as a unit, columnheaders
subsequent to the first block were incorrectly treated as data.

Now we reinitialize the columnheader processing at the start
of each new data block in the file.

Bug #2538