Extend discrete gradient palette (DGP) implementation to postscript terminal
Postscript terminals were excluded from the colorbox rendering changes
introduced in commits 4ba95e0e 29f60657 5989ba41 685c7a77
The reason is that the colorbox rendering routine for postscript terminals
(draw_inside_color_smooth_box_postscript) was implemented separately from
the other terminals.

In this patch, the same drawing routine is used for discrete gradient
palettes as for other terminals, even for postscript terminals.
For smooth gradient palettes (and mixed gradient palettes),
the optimized drawing routines are used as before.

The maxcolors setting is disabled in the output postscript file because
it is meaningless for the discrete gradient palette to limit the colors
by maxcolors. This behavior is the same for other terminals.