user-callable function split( "string"{, "sep"} )
Splits a string into an array of strings.  The cut points are determined
by matching the character sequence in the optional second parameter "sep".
If sep is omitted, is the null string, or is the single character " ",
then any number of adjacent whitespace characters is treated as a field
separator.  Otherwise each field is terminated by the exact character
sequence in sep.

The three examples below each produce an array [ "A", "B", "C", "D" ]
    t1 = split( "A B C D" )
    t2 = split( "A B C D", " ")
    t3 = split( "A;B;C;D", ";")
However the command
    t4 = split( "A;B; C;D", "; " )
produces an array containing only two strings [ "A;B", "C;D" ] because
the two-character separator sequence "; " is found only once.

Realistic use:
    # Extract base names of data files
    files = split( "`ls -1 *.dat`"."\n", ".dat\n" )
    nfiles = |files|
    do for (i = 1:nfiles) {
	... something with base name ...
    }