CHANGE: "set tics {front|back}" does not re-enable tics that are off
Previously "set tics" with any of the options front/back/scale
acted like "set tics" in that it reset all of the axes to the default
on/off/axis/border/mirror settings.

E.g.
	unset xtics; unset ytics
        set rtics
	set tics front

had the undesired side effect of re-enabling x and y tics.
Now it doesn't.

We still re-enable tics on "set tics scale", however, because
"set tics scale 0.0" is sometimes used for the side effect of
extending an auto-scaled range to the next tic even though
the next tic is not visible.