Fix autoscaled log r
Autoscaling the minimum bound on a logscaled r axis cannot work,
because the transformed log(rmin) needs to be known before hand
in order to process data points as they are input.
Now "set log r" will force rmin = 0.1 and turn off AUTOSCALE_MIN.

Autoscaling the maximum bound on logscale r is possible but it was broken.
This fixes it by updating both the primary and secondary maximum
in parallel during input.