"set polar grid" and plot "with surface" produce a polar heatmap
Syntax:
      set polar grid {<theta_segments>, <radial_segments>}
                     { qnorm {<power>} | gauss | cauchy | exp | box | hann }
                     { kdensity } { scale <scale> }
                     {theta [min:max]} {r [min:max]}

The polar grid settings are used in conjunction with the plot style
`with surface` to generate a heat map from a set of polar coordinate points.
The surface consists of a grid filling a circle divided into segments formed
by discrete ranges on theta and r.

Each segment is assigned a value derived from the input set of individual
scattered points [x,y,z] by applying a filter operation.
The default filter is `qnorm 1`, which averages each point's z value weighted
by the inverse of the point's distance from the center of that grid segment.

Alternative filter operations gauss, cauchy, exp, box, and hann act as they
do for "set dgrid3d".

`kdensity`: This keyword tells the program to use the weighted sum of
contributions from all points rather than the weighted average.

`scale`: This scale facter (default 1.0) is applied to all distances prior to
using them in the weighting calculation.

Masking:  All input points are used to calculate grid values.
The full gridded surface always spans theta range [0:360] and the radial
defined by autoscaling or by a previous `set rrange` command.
However the portion of the surface that actually appears in the plot
can be restricted to a truncated wedge bounded by lower and upper
limits on theta and r.  Theta limits must be given in degrees.

Example: The following command sequence will generate a plot which is
auto-scaled in size to show all input points.
The contributions of all input points are summed, not averaged (`kdensity`),
and only a truncated wedge of the resulting gridded surface is displayed.

     set rrange [0:*]
     set polar grid qnorm kdensity theta [45:136] r [10:50]
     plot DATA with surface, DATA with points