Fix regression: autopromotion of string to int failed on leadin 0
Commit 1128546b replaced atoll with strtoll, with the intent that
it would handle hexadecimal as well as decimal values.

In retrospect it was a bad idea to treat strings with a leading
0 as always meaning "octal", because
(1) In contexts like extracting the numerical part of something
    like a filename data_009, data_010, etc the leading zeros are
    just placeholders in a decimal value.
(2) 08 and 09 failed with the confusing error message
    "Trailing characters after numeric" rather than something
    like "illegal character in octal value".
(3) Real world scripts broke, including the epi_data demo.

This regression affected version 5.4.4 only.

We now partially revert commit 1128546b.
Autopromotion now treats a string as hexadecimal if it begins with 0x,
and explicitly decimal otherwise.  Leading 0s in the absence of 0x
do not signify an octal value.
Furthermore the "Trailing characters after numeric" message is
given as a warning, not a fatal error.

Bugs #2527 #2551