new variable type FUNCTIONBLOCK
New command
    function $myfunction << EOD
       ... body of function block ...
    EOD

Creates and initializes a new function block.
The only syntax difference between declaring a data block
and a function block is the type identifier.

The plan is that just as data blocks are accepted in place of a file
for plot, print, fit, etc, function blocks will be accepted as a
source of gnuplot commands.

Why is this needed when we already allow "load $datablock"
and "call $datablock"?  The intent is that unlike load and call,
the commands in a function block can be executed directly from
the execution stack and yield a return value
    foo = $myfunction(x)

The return and evaluation bits will come next.