New command "return <expression>"
"return <expression>" acts like "exit <expression>" except that
the return value is passed to f_eval() via the global eval_return_value
so that it can be pushed onto the evaluation stack after execution
of a function block.

If the "return" statement is encountered outside a function block
it still acts like "exit" but the return value is lost.