Track names of dummy parameters used in defining a functionblock
Prior to this commit, parameters passed to a functionblock could
only be accessed using the ARGC+ARGV mechanism.  This commit adds
a tracking mechanism for dummy parameter names.
When the functionblock definition specifies dummy parameters,
the parameter names are used to execute code equivalent to:
    function $myfuncblock( a, b, c )  << EOF
      local a = ARGV[1]
      local b = ARGV[2]
      local c = ARGV[3]
      ... body of block ...
    EOF

A FUNCTIONBLOCK value now looks like this:
    struct value
	.type = FUNCTIONBLOCK
	.v = struct fblock {
		.data_array
		.parnames
	     }

TODO:
    - What to do if there is a mismatch between the number of parameters
      the function definition and the number actually passed?
    - 'print $func' and/or 'show var' should list dummy parameters