Do not deliberately leave undefined/NaN values on an otherwise empty stack
Perhaps due to vestigial paranoia about floating-point error handling
the expression evaluation code in evaluate_at() tested for an undefined
result and chose not to pop it from the stack before return, instead
setting an "undefined" flag. This left the bad/undefined/NaN value
on what should have been an empty stack.  It then also skipped checking
the state of the stack presumably because it was inevitably going to
report that the stack was not empty.  I can see no point to this.

Now we pop the value, which may contain NaN but so what?
In an excess of residual paranoia we first check that there will not be
stack underflow.  check_stack() is called but is expected to succeed.
The "undefined" flag is still set.

The motivation for this change is to reduce the number of places where
the stack must be forcibly reset to empty.