Suppress resetting the evaluation stack during function block evaluation
Prior to the introduction of function blocks, the evaluation stack was reset
to empty prior to each and every expression evaluation.
Since function blocks are called from the stack and can then initiate new
expression evaluation, this would mean that anything on the stack prior
to the function block would be gone when the block exits.

Now we introduce a flag evaluate_inside_functionblock that holds
the recursion depth at which function block evaluation was entered.
While it is set, the stack is only reset during error handling.
When the recursion depth returns to that initial level the flag
is cleared.

Note that the flag is always clear at the top level,
and is never set if function blocks are not used.

This change is conservative.  I think it would be safe to skip
the stack reset every time in any case, but this way the behavior
remains the same as before if function blocks are not active.

We now call reset_stack() explicitly during a "reset" command and
during error handling.