Add parameter "clear" to internal routine get_position_default()
get_position_default() is primarily used from "set" commands.
Previously if a 2nd or 3rd coordinate was omitted from the "set" command,
it was always set to zero.  For example
	set label 1 at graph 0.5, 0.1
	set label 1 at graph 0.4
would leave the y coordinate of the label at graph 0.

Setting clear to TRUE retains this behavior. Setting it to FALSE
causes any previous coordinate to be kept, so in the above example
the label would end up at 0.4, 0.1

At this point I only set it to FALSE for "set colorbox {origin|size}"
and "set key offset" but probably there are other commands that would be
more intuitive this way.