local arrays need special treatment
If "local array A[xx]" is treated as a completely normal array
declaration then it leaks memory since the storage is not freed
when variable name A goes out of scope.
But marking it TEMP_ARRAY is not correct either, as the storage might
then be freed too early when the array is used in its proper scope.

This commit adds a new flag type LOCAL_ARRAY that indicates the storage
should be freed only by routine lf_exit_scope().