sharpen:  improve clipping of sharpened line plots
The "sharpen" option can cause the y value of individual points
to reach extremely large positive or negative values.
If the resulting value exceeded INT_MAX the point was treated as
UNDEFINED rather than OUTRANGE, which would cause a gap in the plot
rather than a spike.

This commit introduces a new function map_ysharp which catches this
overflow case, and revises clip_line() to handle the full range
of possible results for (y1 - y0).