info:  Fix navigation to "Bugs" section of docs.
A spurious menu with a "Bugs" menu item is generated by doc2texi.el
at the end of Section 4 (just before Section 5 "Bugs").
This is incorrect and breaks navigation in "info": going backward from
Section 5 sends to Section 5.2 instead of the end of Section 4.
I've attached a patch that removes the 3 lines responsible.