wxt: The export-to-file widget should preserve line properties
The export-to-file widget in the toolbar of the wxt terminal allows
you to save the current plot as a png, svg, or pdf file.
However the linecap and linejoin properties currently active in
the main plot window were not replicated in the svg or pdf output.
It is not clear to me exactly why or when the active settings are
lost, but now we reapply them before exporting the plot.