deprecate "smooth convexhull" in favor of "convexhull smooth path"
This reduces the number of special case tests, particularly because
it avoids the issue of introducing introducing parallel options
"smooth concavehull" or "smooth chishape", etc., as new filters are
added.

For now the old syntax (added July 2021, development version only)
still works but maybe remove it in the version 6 release.