Remove spurious Texinfo comment from gnuplot.info
The docs/doc2texi.el converter transforms the line starting
with "<3" into a Texinfo comment by inserting "@c " at the
beginning of the line (function d2t-get-terminals). However,
the input has not been converted to the Texinfo format yet,
so that function d2t-braces-atsigns, which is called later,
assumes that "@c" is real text and adds a @ in front of it.
As a consequence, one gets a line
  @@c <3 -- all terminal stuff is pulled from the .trm files
in the generated Texinfo file gnuplot.texi, and this line
will be rendered as
  @c <3 -- all terminal stuff is pulled from the .trm files
in the Info manual.
This patch makes doc2texi.el remove the "<3" line instead of
attempting to comment it out.