default fillstyle for circle/polygon/ellipse objects is now "fs empty border"
If you set a linecolor for an empty circle/polygon/ellipse this has the
effect of changing the border property.  I.e.
	set obj 1 fs empty
	set obj 1 lc "red"
is now equivalent to
	set obj 1 fs empty border lc "red"

This is a CHANGE.  Previously the default was "fs solid border -1 fc bgnd"
for circles and ellipses, "fc 'black'" for polygons.

Rectangles remain a separate case and are unchanged by this commit.
Rectangles have always been treated differently than other object types.