remove option ./configure --with-tutorial
The LaTeX tutorial was based on old latex terminals that are no longer
supported. It was functionally replaced by a latex terminal demo
demo/latex_demo.tex.

The tutorial was already removed as a top level build target
by commit f623755.  Now it is not listed as a configure option.