bugfix: border color for filled curves and polygons with variable fillcolor
The filled area of a filled curve is drawn first.
If the fillstyle has a border a separate call to plot_lines() is then made.
Polygons are the same.
But plot_lines() checks for variable color again and gets the fill color not
the border color.  Disable this check for plot styles filledcurves and polygons.