Treat set_color(NaN) as LT_NODRAW rather than LT_BACKGROUND
That is, NaN now triggers term->linetype(LT_NODRAW)
rather than term->set_color(BACKGROUND_COLOR).

I think this is clearly better if you are actually drawing lines.
If you are using it for solid fill I think the result is
terminal-dependent.  Some will not draw anything (good) and others
might use the "color" of LT_NODRAW which is background color
(not as good but the same as it was doing before).